+++ api/src/adapters/aws/sesv2_adapter.py
